\name{dupSdf}
\alias{dupSdf}
\title{ Duplicate SDF }
\description{ Duplicates an SDF.  }
\usage{
dupSdf(sdf)
}
\arguments{
  \item{sdf}{ The SDF object to be duplicated. }
}
\value{
An sqlite.data.frame that is a duplicate of the argument. The internal
name will be autogenerated, i.e. the default data<n>.
}
\author{Miguel A. R. Manese}
\seealso{ 
    \code{\link[SQLiteDF]{renameSdf}}
}
\keyword{data}
