\name{SQLiteDF-package}
\alias{SQLiteDF-package}
\alias{SQLiteDF}
\docType{package}
\title{
Utilities to store data into SQLite data frames.
}
\description{
Not much really.
}
\details{
\tabular{ll}{
Package: \tab SQLiteDF\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2006-06-09\cr
License: \tab What license is it under?\cr
}
}
\author{
Miguel A. R. Manese
Maintainer: Miguel A. R. Manese <jjonphl@gmail.com>
}
\references{
NA
}
\keyword{ package }
\seealso{
}
\examples{
}
