\name{SQLiteDF-package}
\alias{SQLiteDF-package}
\alias{SQLiteDF}
\docType{package}
\title{ SQLite data frames }
\description{ S3 and utility methods to implement SQLite data frames. }
\details{
This package aims to transparently operate on data frames stored in
SQLite databases.
}
\author{
Miguel A. R. Manese
Maintainer: Miguel A. R. Manese <jjonphl@gmail.com>
}
\keyword{ package }
